All stats are now correctly displayed on Menu screen. This was done with the help of helper functions which copied the functions/variables in now deleted Stats data class.
